requirements.txt
aiofiles==24.1.0
anyio==4.6.2.post1
argcomplete==3.5.1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.55
botocore==1.35.55
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
dacite==1.7.0
Deprecated==1.2.14
dill==0.3.9
dnspython==2.7.0
fiftyone==1.0.1
fiftyone-brain==0.17.0
fiftyone_db==1.1.7
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
glob2==0.7
graphql-core==3.2.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
imageio==2.36.0
inflate64==1.0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kaleido==0.2.1
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mongoengine==0.24.2
motor==3.5.3
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pprintpp==0.4.0
priority==2.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.21.0
pydash==8.0.4
pymongo==4.8.0
pyparsing==3.2.0
pyppmd==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
google-maps-places
pyzstd==0.16.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.41.2
strawberry-graphql==0.248.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
universal-analytics-python3==1.1.1
urllib3==2.2.3
voxel51-eta==0.13.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.14.2
